move tx status update logs to consensus
