+++ backend/database_handler/transactions_processor.py
@@ -7,22 +7,14 @@ from sqlalchemy.orm import Session
 from .models import TransactionStatus
 from eth_utils import to_bytes, keccak, is_address
 import json
-from backend.protocol_rpc.message_handler.base import MessageHandler
-from backend.protocol_rpc.message_handler.types import (
-    LogEvent,
-    EventType,
-    EventScope,
-)
 
 
 class TransactionsProcessor:
     def __init__(
         self,
         session: Session,
-        msg_handler: MessageHandler = None,
     ):
         self.session = session
-        self.msg_handler = msg_handler
 
     @staticmethod
     def _parse_transaction_data(transaction_data: Transactions) -> dict:
@@ -151,7 +143,6 @@ class TransactionsProcessor:
 
         transaction.status = new_status
         self.session.commit()
-        self.dispatch_transaction_status_update(transaction_hash, new_status)
 
     def set_transaction_result(self, transaction_hash: str, consensus_data: dict):
         transaction = (
@@ -162,23 +153,3 @@ class TransactionsProcessor:
         transaction.consensus_data = consensus_data
 
         self.session.commit()
-        self.dispatch_transaction_status_update(
-            transaction_hash, TransactionStatus.FINALIZED
-        )
-
-    def dispatch_transaction_status_update(
-        self, transaction_hash: str, new_status: TransactionStatus
-    ):
-        if self.msg_handler:
-            self.msg_handler.send_message(
-                LogEvent(
-                    "transaction_status_updated",
-                    EventType.INFO,
-                    EventScope.CONSENSUS,
-                    f"{str(new_status.value)} {str(transaction_hash)}",
-                    {
-                        "hash": str(transaction_hash),
-                        "new_status": str(new_status.value),
-                    },
-                )
-            )
